+++ src/eh0/spi.rs
@@ -182,7 +182,7 @@ impl spi::Transfer<u8> for Mock {
         );
         assert_eq!(
             &w.expected_data, &buffer,
-            "spi::write data does not match expectation"
+            "spi::transfer write data does not match expectation"
         );
         assert_eq!(
             buffer.len(),
@@ -323,89 +323,46 @@ mod test {
     }
 
     #[test]
-    #[should_panic(
-        expected = "assertion failed: `(left == right)`\n  left: `[10, 12]`,\n right: `[10, 12, 12]`: spi::write data does not match expectation"
-    )]
+    #[should_panic(expected = "spi::write data does not match expectation")]
     fn test_spi_mock_write_err() {
         let expectations = [Transaction::write(vec![10, 12])];
         let mut spi = Mock::new(&expectations);
-
         spi.write(&vec![10, 12, 12]).unwrap();
-
-        spi.done();
     }
 
     #[test]
-    #[should_panic(
-        expected = "assertion failed: `(left == right)`\n  left: `[10, 12]`,\n right: `[10, 12, 12]`: spi::write_iter data does not match expectation"
-    )]
+    #[should_panic(expected = "spi::write_iter data does not match expectation")]
     fn test_spi_mock_write_iter_err() {
         let expectations = [Transaction::write(vec![10, 12])];
         let mut spi = Mock::new(&expectations);
-
         spi.write_iter(vec![10, 12, 12u8]).unwrap();
-
-        spi.done();
     }
 
     #[test]
-    #[should_panic(
-        expected = "assertion failed: `(left == right)`\n  left: `[12, 15]`,\n right: `[12, 13]`"
-    )]
+    #[should_panic(expected = "spi::transfer write data does not match expectation")]
     fn test_spi_mock_transfer_err() {
         let expectations = [Transaction::transfer(vec![10, 12], vec![12, 15])];
         let mut spi = Mock::new(&expectations);
-
-        let mut v = vec![10, 12];
-        spi.transfer(&mut v).unwrap();
-
-        assert_eq!(v, vec![12, 13]);
-
-        spi.done();
+        spi.transfer(&mut vec![10, 13]).unwrap();
     }
 
     #[test]
-    #[should_panic(
-        expected = "assertion failed: `(left == right)`\n  left: `[1, 2]`,\n right: `[10, 12]`: spi::write data does not match expectation"
-    )]
-    fn test_spi_mock_transfer_response_err() {
-        let expectations = [Transaction::transfer(vec![1, 2], vec![3, 4, 5])];
-        let mut spi = Mock::new(&expectations);
-
-        let mut v = vec![10, 12];
-        spi.transfer(&mut v).unwrap();
-
-        assert_eq!(v, vec![12, 13]);
-
-        spi.done();
-    }
-
-    #[test]
-    #[should_panic(
-        expected = "assertion failed: `(left == right)`\n  left: `Transfer`,\n right: `Write`: spi::write unexpected mode"
-    )]
-    fn test_spi_mock_mode_err() {
-        let expectations = [Transaction::transfer(vec![10, 12], vec![])];
-        let mut spi = Mock::new(&expectations);
-
-        spi.write(&vec![10, 12, 12]).unwrap();
-
-        spi.done();
-    }
-
-    #[test]
-    #[should_panic(
-        expected = "assertion failed: `(left == right)`\n  left: `[10, 12]`,\n right: `[10, 12, 12]`: spi::write data does not match expectation"
-    )]
+    #[should_panic(expected = "spi::write data does not match expectation")]
     fn test_spi_mock_multiple_transaction_err() {
         let expectations = [
             Transaction::write(vec![10, 12]),
             Transaction::write(vec![10, 12]),
         ];
         let mut spi = Mock::new(&expectations);
+        spi.write(&vec![10, 12, 10]).unwrap();
+    }
 
+    #[test]
+    #[should_panic(expected = "spi::write unexpected mode")]
+    fn test_spi_mock_mode_err() {
+        let expectations = [Transaction::transfer(vec![10, 12], vec![])];
+        let mut spi = Mock::new(&expectations);
+        // Write instead of transfer
         spi.write(&vec![10, 12, 12]).unwrap();
-
-        spi.done();
     }
 }